stream_ui_updates: Cut import of stream_settings_ui.
